+++ Informatica/src/model/Monitor.java
@@ -19,7 +19,7 @@ public class Monitor extends Equipamento{
         super();
     }
 
-    public Monitor(String tipo, String tamanhoTela, String resolucao, int codEquipamento, String modelo, String marca, int quantEstoque, String categoria, double valorDiaria) {
+    public Monitor(String tipo, String tamanhoTela, String resolucao, int codEquipamento, String modelo, String marca, int quantEstoque, String categoria, float valorDiaria) {
         super(codEquipamento, modelo, marca, quantEstoque, categoria, valorDiaria);
         this.tipo = tipo;
         this.tamanhoTela = tamanhoTela;
